Run tests against Review App on each pull request
Prior to this commit the tests were always running against the web app's
prouduction URL.  Now the tests are triggered only on pull requests and
run against the [Heroku Review App][1] that is automatically created on
the creation of every pull request.

In a future PR we will want to have the functional test running on every
push to master too (e.g. after the PR is merged).  That could be done by
invoking the Heroku API to create a Review App on the fly on each push
to the master branch, for instance.  Then if the tests fail we would not
deploy to prouduction.

[1]: https://devcenter.heroku.com/articles/github-integration-review-apps