Use Prism mock instead of real API
One interesting thing about this is that when we used Prism to mock the
API call, it instantly flagged that we were missing the required
authorization header in our request (we ran Prism with the `--errors`
flag which is something we should always do, to discover errors like
this).

This switch to use the Prism mock simply involved changing the
`PETSTORE_URL` environment variable to point to Prism, i.e.
`http://127.0.0.1:4010/` when running locally, and setting up a
[Stoplight hosted Prism mock][1] when running on a Review App, i.e.
`https://stoplight.io/mocks/agilepathway/java-openapi-provider/84330/`.

In a later commit we'll setup the URL to be dynamic based on the Git
branch, so that we can use an individual Prism mock server for every
branch.

[1]: https://meta.stoplight.io/docs/platform/ZG9jOjIwMTI4MA-working-with-mock-servers