Use branch-specific mock server on review apps (#33)
Prior to this commit the Prism mock server that is used on our pull
request triggered [Review Apps][1] was always based on the trunk branch
of the [provider repo][2] (which has the OpenApi spec that the Prism
mock server is based on).  Now in our pull request CI/CD pipeline we
will use a Prism mock server based on the matching git branch of the
provider repo, defaulting to the trunk repo if there is no matching git
branch.

This means that when the consumer wants to request a change to the
provider's OpenAPI spec (through our
[consumer-driven-contract-testing][3] approach), the consumer can go
ahead and create a branch on the provider app's repo with the change -
then allowing the consumer to run their tests against the proposed
change on the Review App.  So in short this is a big enabler for a real
consumer-driven approach.

[1]: https://devcenter.heroku.com/articles/github-integration-review-apps
[2]: https://github.com/agilepathway/java-openapi-provider
[3]: https://martinfowler.com/articles/consumerDrivenContracts.html