Tag trunk commits with contract repo HEAD commit
The CI/CD pipeline now tags any commit to trunk (i.e. any merge to the
trunk branch) (which has successfully run the post-merge contract tests)
with the HEAD commit SHA of the contract repo.  This is so that we can
check that the provider in any of our integrated environments is
compatible with the contract that the consumer needs, before deploying
to that integrated environment.

For example, if we have amended the contract with a new feature but the
provider has not yet added their implementation of this new feature to
production, then we will be able to set up our CI/CD (in a future PR to
come after this one) to fail the build (by running the contract tests
for the contract demanded by the consumer against a mock provider using
the provider commit that is in production) and therefore correctly
prevent us deploying the consumer to production.

When we tag the trunk commit we append the date and time to the end of
the tag.  This is because tags have to be unique, and multiple commits
can be linked to the same contract repo commit.