Add CI "can I deploy" check before deploying
This commit adds a job to the CI pipeline which checks that it is safe
to deploy to our integrated environments (pre-production and production)
before deploying.  The check is done by running the contract tests on
the specific commit on the contract repo that is associated (tagged)
with our consumer commit that we are trying to deploy.  The contract
tests run against the real provider, [using Prism as a validation
proxy][1].  We will enhance this functionality further in the future
by spinning up an ephemeral environment for the provider rather than
using the real production provider.

[1]: https://meta.stoplight.io/docs/prism/docs/guides/03-validation-proxy.md