Perform `can-i-deploy` on ephemeral environment (#53)
Prior to this commit the `can-i-deploy` checks were running against the
real integrated environment, which is not what we want.  Now we use our
`ephemeral-petstore-service` Heroku environment (which isn't really
ephemeral, but serves the purpose by highlighting that it would be
ephemeral in a more "productionised" example).  We could experiment
in a later commit with using Heroku's [Review App API][1] to create a
genuinely ephemeral environment, but that's for a later date (post-MVP).

[1]: https://devcenter.heroku.com/articles/github-integration-review-apps#review-apps-api